New ThreeWayHandshake to be beta tested
Tested internally in EUT in LAN settings with no vpn involved. Distinguished client and server which exchanged seq_id msg to wait one another at the same sync point with a SYNC -> SYNC-ACK -> ACK pattern to agree that they are on the same stage.
More details provided when reaching the final test stage of this version. Further improvements:
-  avoiding sending useless msgs
- check reset of seq_id when delay among the two parties are really high